code: relocate adjust_for_inflation to _helpers.py
