+++ scripts/compile_cost_assumptions_usa.py
@@ -13,8 +13,8 @@ import pathlib
 
 import numpy as np
 import pandas as pd
-from _helpers import mock_snakemake
-from compile_cost_assumptions import carbon_flow, prepare_inflation_rate
+from _helpers import mock_snakemake, adjust_for_inflation
+from compile_cost_assumptions import prepare_inflation_rate
 
 
 def get_conversion_dictionary(flag: str):
@@ -239,45 +239,6 @@ def calculate_fom_percentage(x, dataframe, columns_list):
 def replace_value_name(dataframe, conversion_dict, column_name):
     dataframe[column_name] = dataframe[column_name].replace(conversion_dict)
     return dataframe
-
-
-def adjust_for_inflation(inflation_rate, costs, techs, eur_year, col):
-    """
-    Adjust the investment costs for the specified techs for inflation.
-
-    techs: str or list
-        One or more techs in costs index for which the inflation adjustment is done.
-    ref_year: int
-        Reference year for which the costs are provided and based on which the inflation adjustment is done.
-    costs: Dataframe containing the costs data with multiindex on technology and one index key 'investment'.
-    """
-
-    def get_factor(inflation_rate, ref_year, eur_year):
-        if (pd.isna(ref_year)) or (ref_year < 1900):
-            return np.nan
-        if ref_year == eur_year:
-            return 1
-        mean = inflation_rate.mean()
-        if ref_year < eur_year:
-            new_index = np.arange(ref_year + 1, eur_year + 1)
-            df = 1 + inflation_rate.reindex(new_index).fillna(mean)
-            return df.cumprod().loc[eur_year]
-        else:
-            new_index = np.arange(eur_year + 1, ref_year + 1)
-            df = 1 + inflation_rate.reindex(new_index).fillna(mean)
-            return 1 / df.cumprod().loc[ref_year]
-
-    inflation = costs.currency_year.apply(
-        lambda x: get_factor(inflation_rate, x, eur_year)
-    )
-
-    paras = ["investment", "VOM", "fuel"]
-    filter_i = costs.technology.isin(techs) & costs.parameter.isin(paras)
-    costs.loc[filter_i, col] = costs.loc[filter_i, col].mul(
-        inflation.loc[filter_i], axis=0
-    )
-
-    return costs
 
 
 def pre_process_manual_input_usa(
